Fix error delete user

+++ app/Services/UserService.php
@@ -65,6 +65,10 @@ class UserService
 
     public function deleteUser($id)
     {
+        if (auth()->id() == $id) {
+            throw new \Exception('Bạn không thể xóa tài khoản của chính mình.');
+        }
+        
         return $this->userRepository->delete($id);
     }
 }
